@@ -1,16 +1,15 @@
 ﻿using System;
 
-namespace pdbget
-{
-    [AttributeUsage(AttributeTargets.Field, Inherited = false, AllowMultiple = false)]
-    public sealed class NameAttribute : Attribute
-    {
-        public NameAttribute(string name)
-        {
-            Name = name;
-        }
+namespace pdbget;
 
-        public string Name { get; }
-        public string? InternalName { get; set; }
+[AttributeUsage(AttributeTargets.Field, Inherited = false, AllowMultiple = false)]
+public sealed class NameAttribute : Attribute
+{
+    public NameAttribute(string name)
+    {
+        Name = name;
     }
+
+    public string Name { get; }
+    public string? InternalName { get; set; }
 }
