@@ -1,27 +1,26 @@
 ﻿using pdbget.Helpers;
 
-namespace pdbget
+namespace pdbget;
+
+public static class AminoAcidExtensions
 {
-    public static class AminoAcidExtensions
+    public static AminoAcid ParseAminoAcid(this string s)
     {
-        public static AminoAcid ParseAminoAcid(this string s)
-        {
-            return EnumByMatchParser<AminoAcid>.Parse(s);
-        }
+        return EnumByMatchParser<AminoAcid>.Parse(s);
+    }
 
-        public static bool TryParseAminoAcid(this string s, out AminoAcid value)
-        {
-            return EnumByMatchParser<AminoAcid>.TryParse(s, out value);
-        }
+    public static bool TryParseAminoAcid(this string s, out AminoAcid value)
+    {
+        return EnumByMatchParser<AminoAcid>.TryParse(s, out value);
+    }
 
-        public static string GetName(this AminoAcid value)
-        {
-            return EnumAnnotationHelper<AminoAcid>.GetAttribute<NameAttribute>(value).Name;
-        }
+    public static string GetName(this AminoAcid value)
+    {
+        return EnumAnnotationHelper<AminoAcid>.GetAttribute<NameAttribute>(value).Name;
+    }
 
-        public static char GetCode(this AminoAcid value)
-        {
-            return EnumAnnotationHelper<AminoAcid>.GetAttribute<NameAttribute>(value).InternalName![0];
-        }
+    public static char GetCode(this AminoAcid value)
+    {
+        return EnumAnnotationHelper<AminoAcid>.GetAttribute<NameAttribute>(value).InternalName![0];
     }
 }
