@@ -1,67 +1,64 @@
-﻿using System;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using System.Text;
-using System.Threading;
 
-namespace pdbget.Helpers
+namespace pdbget.Helpers;
+
+internal static class ShellHelper
 {
-    internal static class ShellHelper
+    public static (string program, string arguments) SplitCommandLine(this string commandLine)
     {
-        public static (string program, string arguments) SplitCommandLine(this string commandLine)
+        int cmdBegin = 0, cmdEnd = commandLine.Length;
+        if (commandLine[0] == '"' || commandLine[0] == '\'')
         {
-            int cmdBegin = 0, cmdEnd = commandLine.Length;
-            if (commandLine[0] == '"' || commandLine[0] == '\'')
+            int idx = commandLine.IndexOf(commandLine[0], 1);
+            if (idx != -1)
             {
-                int idx = commandLine.IndexOf(commandLine[0], 1);
-                if (idx != -1)
-                {
-                    cmdBegin = 1;
-                    cmdEnd = idx;
-                }
+                cmdBegin = 1;
+                cmdEnd = idx;
             }
-            else
+        }
+        else
+        {
+            int idx = commandLine.IndexOf(' ');
+            if (idx != -1)
             {
-                int idx = commandLine.IndexOf(' ');
-                if (idx != -1)
-                {
-                    cmdBegin = 0;
-                    cmdEnd = idx;
-                }
+                cmdBegin = 0;
+                cmdEnd = idx;
             }
-
-            return (commandLine[cmdBegin..cmdEnd], commandLine[(cmdEnd + 1)..].TrimStart());
         }
 
-        public static void RunCommand(this string program, string arguments, out string stdout, out string stderr)
+        return (commandLine[cmdBegin..cmdEnd], commandLine[(cmdEnd + 1)..].TrimStart());
+    }
+
+    public static void RunCommand(this string program, string arguments, out string stdout, out string stderr)
+    {
+        // Create your Process
+        using Process process = new();
+        process.StartInfo.FileName = program;
+        process.StartInfo.Arguments = arguments;
+        process.StartInfo.UseShellExecute = false;
+        process.StartInfo.CreateNoWindow = true;
+        process.StartInfo.RedirectStandardOutput = true;
+        process.StartInfo.RedirectStandardError = true;
+
+        StringBuilder sberr = new();
+
+        // Set ONLY ONE handler here.
+        process.ErrorDataReceived += (sender, e) =>
         {
-            // Create your Process
-            using var process = new Process();
-            process.StartInfo.FileName = program;
-            process.StartInfo.Arguments = arguments;
-            process.StartInfo.UseShellExecute = false;
-            process.StartInfo.CreateNoWindow = true;
-            process.StartInfo.RedirectStandardOutput = true;
-            process.StartInfo.RedirectStandardError = true;
+            sberr.AppendLine(e.Data);
+        };
 
-            var sberr = new StringBuilder();
+        // Start process
+        process.Start();
 
-            // Set ONLY ONE handler here.
-            process.ErrorDataReceived += (sender, e) =>
-            {
-                sberr.AppendLine(e.Data);
-            };
+        // Read one element asynchronously
+        process.BeginErrorReadLine();
 
-            // Start process
-            process.Start();
+        // Read the other one synchronously
+        stdout = process.StandardOutput.ReadToEnd();
 
-            // Read one element asynchronously
-            process.BeginErrorReadLine();
-
-            // Read the other one synchronously
-            stdout = process.StandardOutput.ReadToEnd();
-
-            process.WaitForExit();
-            stderr = sberr.ToString();
-        }
+        process.WaitForExit();
+        stderr = sberr.ToString();
     }
 }
