@@ -1,7 +1,6 @@
-﻿namespace pdbget.PdbFormat
+﻿namespace pdbget.PdbFormat;
+
+public abstract class AbstractChainedRecord : Record
 {
-    public abstract class AbstractChainedRecord : Record
-    {
-        public char ChainId { get; set; }
-    }
+    public char ChainId { get; set; }
 }
