@@ -1,17 +1,16 @@
 ﻿using pdbget.Helpers;
 
-namespace pdbget.PdbFormat
-{
-    public static class RecordTypeExtensions
-    {
-        public static RecordType ParseRecordType(this string s)
-        {
-            return EnumByMatchParser<RecordType>.Parse(s);
-        }
+namespace pdbget.PdbFormat;
 
-        public static bool TryParseRecordType(this string s, out RecordType value)
-        {
-            return EnumByMatchParser<RecordType>.TryParse(s, out value);
-        }
+public static class RecordTypeExtensions
+{
+    public static RecordType ParseRecordType(this string s)
+    {
+        return EnumByMatchParser<RecordType>.Parse(s);
+    }
+
+    public static bool TryParseRecordType(this string s, out RecordType value)
+    {
+        return EnumByMatchParser<RecordType>.TryParse(s, out value);
     }
 }
